+++ app/Filament/Resources/RedisConfigurationResource/Pages/ListRedisConfigurations.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\RedisConfigurationResource\Pages;
+
+use App\Filament\Resources\RedisConfigurationResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListRedisConfigurations extends ListRecords
+{
+    protected static string $resource = RedisConfigurationResource::class;
+}