+++ app/Filament/Resources/RedisConfigurationResource/Pages/ViewRedisConfiguration.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\RedisConfigurationResource\Pages;
+
+use App\Filament\Resources\RedisConfigurationResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewRedisConfiguration extends ViewRecord
+{
+    protected static string $resource = RedisConfigurationResource::class;
+}